fix(iterate): split the silent-revert detector, and give the test repos an identity
Two follow-ups from the gates on this run's own PR.

The bloat Stop hook caught silent_revert.py at 305 lines with no baseline entry
— a new crossing. Split rather than baselined: the merge walk (_integration_merges
+ dropped_lines) moves to silent_revert_detect.py, leaving the check module to
turn a set of dropped lines into a CheckResult. That is the seam the three
earlier splits left behind — everything in the new module walks the repository,
nothing in it decides what the operator is told. 305 -> 174 + 159.

CI caught test_a_resolvable_local_ref_is_still_preferred_when_it_is_ahead: a
clone does not inherit the origin's LOCAL git identity and a CI runner has no
global one, so `git commit` failed there, main never moved ahead, and the
assertion measured an unmoved ref. Green on a dev box, red on CI. The clone now
gets its own identity, and the test asserts its SETUP before its subject so the
same silence cannot pass again. Verified with GIT_CONFIG_GLOBAL=/dev/null.

Run-ID: iterate-2026-07-28-f11-verifies-own-run
